+++ Kernel/System/CheckItem.pm
@@ -118,7 +118,10 @@ sub CheckEmail {
 
     # check needed stuff
     if ( !$Param{Address} ) {
-        $Self->{LogObject}->Log( Priority => 'error', Message => 'Need Address!' );
+        $Self->{LogObject}->Log(
+            Priority => 'error',
+            Message => 'Need Address!'
+        );
         return;
     }
 
@@ -163,30 +166,48 @@ sub CheckEmail {
         my $Resolver = Net::DNS::Resolver->new();
         if ($Resolver) {
 
+            # it's no fun to have this hanging in the web interface
+            $Resolver->tcp_timeout(3);
+            $Resolver->udp_timeout(3);
+
             # check if we need to use a specific name server
             my $Nameserver = $Self->{ConfigObject}->Get('CheckMXRecord::Nameserver');
             if ($Nameserver) {
                 $Resolver->nameservers($Nameserver);
             }
 
-            # A-record lookup
+            # A-record lookup to verify proper DNS setup
             my $Packet = $Resolver->send( $Host, 'A' );
             if ( !$Packet ) {
                 $Self->{ErrorType} = 'InvalidDNS';
                 $Error = "DNS problem: " . $Resolver->errorstring();
                 $Self->{LogObject}->Log(
                     Priority => 'error',
-                    Message  => "DNS problem: " . $Resolver->errorstring(),
+                    Message  => $Error,
                 );
             }
 
             else {
-
+                # RFC 5321: first check MX record and fallback to A record if present.
                 # mx record lookup
                 my @MXRecords = Net::DNS::mx( $Resolver, $Host );
                 if ( !@MXRecords ) {
-                    $Error = "no mail exchanger (mx) found!";
-                    $Self->{ErrorType} = 'InvalidMX';
+                    $Self->{LogObject}->Log(
+                        Priority => 'notice',
+                        Message =>
+                            "$Host has no mail exchanger (MX) defined, trying A resource record instead.",
+                    );
+                    # see if our previous A-record lookup returned a RR
+                    if ( scalar $Packet->answer() eq 0 ) {
+
+                        $Self->{ErrorType} = 'InvalidMX';
+                        $Error = "$Host has no mail exchanger (MX) or A resource record defined.";
+
+                        $Self->{LogObject}->Log(
+                            Priority => 'error',
+                            Message  => $Error,
+                        );
+                    }
                 }
             }
         }